requirements.txt
asgiref==3.11.1
Django==6.0.4
django-phonenumber-field==8.4.0
phonenumbers==9.0.29
sqlparse==0.5.5
tzdata==2026.2
